+++ x-pack/legacy/plugins/ml/public/data_frame_analytics/pages/analytics_management/components/create_analytics_button/create_analytics_button.tsx
@@ -32,7 +32,7 @@ export const CreateAnalyticsButton: FC = () => {
       data-test-subj="mlDataFrameAnalyticsButtonCreate"
     >
       {i18n.translate('xpack.ml.dataframe.analyticsList.createDataFrameAnalyticsButton', {
-        defaultMessage: 'Create data frame analytics job',
+        defaultMessage: 'Create outlier detection job',
       })}
     </EuiButton>
   );
