+++ x-pack/legacy/plugins/ml/public/data_frame_analytics/pages/analytics_management/components/create_analytics_modal/create_analytics_modal.tsx
@@ -35,7 +35,7 @@ export const CreateAnalyticsModal: FC<CreateAnalyticsFormProps> = ({
         <EuiModalHeader>
           <EuiModalHeaderTitle>
             {i18n.translate('xpack.ml.dataframe.analytics.create.modalHeaderTitle', {
-              defaultMessage: 'Create data frame analytics job',
+              defaultMessage: 'Create outlier detection job',
             })}
           </EuiModalHeaderTitle>
         </EuiModalHeader>
