+++ x-pack/legacy/plugins/ml/public/data_frame_analytics/pages/analytics_management/hooks/use_create_analytics_form/use_create_analytics_form.ts
@@ -8,115 +8,19 @@ import { useReducer } from 'react';
 
 import { i18n } from '@kbn/i18n';
 
-import { checkPermission } from '../../../../privilege/check_privilege';
-import { ml } from '../../../../services/ml_api_service';
-import { useKibanaContext } from '../../../../contexts/kibana';
+import { ml } from '../../../../../services/ml_api_service';
+import { useKibanaContext } from '../../../../../contexts/kibana';
 
-import { isValidIndexName } from '../../../../../common/util/es_utils';
+import { useRefreshAnalyticsList, DataFrameAnalyticsOutlierConfig } from '../../../../common';
 
 import {
-  isAnalyticsIdValid,
-  useRefreshAnalyticsList,
-  DataFrameAnalyticsId,
-  DataFrameAnalyticsOutlierConfig,
-} from '../../../common';
-
-export type EsIndexName = string;
-export type IndexPatternTitle = string;
-
-interface RequestMessage {
-  error?: string;
-  message: string;
-}
-
-interface State {
-  createIndexPattern: boolean;
-  destinationIndex: EsIndexName;
-  destinationIndexNameExists: boolean;
-  destinationIndexNameEmpty: boolean;
-  destinationIndexNameValid: boolean;
-  destinationIndexPatternTitleExists: boolean;
-  disabled: boolean;
-  indexNames: EsIndexName[];
-  indexPatternTitles: IndexPatternTitle[];
-  indexPatternsWithNumericFields: IndexPatternTitle[];
-  isJobCreated: boolean;
-  isJobStarted: boolean;
-  isModalButtonDisabled: boolean;
-  isModalVisible: boolean;
-  isValid: boolean;
-  jobId: DataFrameAnalyticsId;
-  jobIds: DataFrameAnalyticsId[];
-  jobIdExists: boolean;
-  jobIdEmpty: boolean;
-  jobIdValid: boolean;
-  requestMessages: RequestMessage[];
-  sourceIndex: EsIndexName;
-  sourceIndexNameExists: boolean;
-  sourceIndexNameEmpty: boolean;
-  sourceIndexNameValid: boolean;
-}
-
-export const getInitialState = (): State => ({
-  createIndexPattern: false,
-  destinationIndex: '',
-  destinationIndexNameExists: false,
-  destinationIndexNameEmpty: true,
-  destinationIndexNameValid: false,
-  destinationIndexPatternTitleExists: false,
-  disabled:
-    !checkPermission('canCreateDataFrameAnalytics') ||
-    !checkPermission('canStartStopDataFrameAnalytics'),
-  indexNames: [],
-  indexPatternTitles: [],
-  indexPatternsWithNumericFields: [],
-  isJobCreated: false,
-  isJobStarted: false,
-  isModalVisible: false,
-  isModalButtonDisabled: false,
-  isValid: false,
-  jobId: '',
-  jobIds: [],
-  jobIdExists: false,
-  jobIdEmpty: true,
-  jobIdValid: false,
-  requestMessages: [],
-  sourceIndex: '',
-  sourceIndexNameExists: false,
-  sourceIndexNameEmpty: true,
-  sourceIndexNameValid: false,
-});
-
-const validate = (state: State): State => {
-  state.isValid =
-    !state.jobIdEmpty &&
-    state.jobIdValid &&
-    !state.jobIdExists &&
-    !state.sourceIndexNameEmpty &&
-    state.sourceIndexNameValid &&
-    !state.destinationIndexNameEmpty &&
-    state.destinationIndexNameValid &&
-    (!state.destinationIndexPatternTitleExists || !state.createIndexPattern);
-
-  return state;
-};
-
-enum ACTION {
-  ADD_REQUEST_MESSAGE = 'add_request_message',
-  RESET_REQUEST_MESSAGES = 'reset_request_messages',
-  CLOSE_MODAL = 'close_modal',
-  OPEN_MODAL = 'open_modal',
-  RESET_FORM = 'reset_form',
-  SET_FORM_STATE = 'set_form_state',
-}
-
-type Action =
-  | { type: ACTION.ADD_REQUEST_MESSAGE; requestMessage: RequestMessage }
-  | { type: ACTION.RESET_REQUEST_MESSAGES }
-  | { type: ACTION.CLOSE_MODAL }
-  | { type: ACTION.OPEN_MODAL }
-  | { type: ACTION.RESET_FORM }
-  | { type: ACTION.SET_FORM_STATE; payload: Partial<State> };
+  getInitialState,
+  reducer,
+  IndexPatternTitle,
+  RequestMessage,
+  State,
+  ACTION,
+} from './reducer';
 
 export interface Actions {
   closeModal: () => void;
@@ -131,82 +35,10 @@ export interface CreateAnalyticsFormProps {
   formState: State;
 }
 
-export function reducer(state: State, action: Action): State {
-  switch (action.type) {
-    case ACTION.ADD_REQUEST_MESSAGE:
-      state.requestMessages.push(action.requestMessage);
-      return state;
-
-    case ACTION.RESET_REQUEST_MESSAGES:
-      return { ...state, requestMessages: [] };
-
-    case ACTION.CLOSE_MODAL:
-      return { ...state, isModalVisible: false };
-
-    case ACTION.OPEN_MODAL:
-      return { ...state, isModalVisible: true };
-
-    case ACTION.RESET_FORM:
-      return getInitialState();
-
-    case ACTION.SET_FORM_STATE:
-      const newState = { ...state, ...action.payload };
-
-      // update state attributes which are derived from other state attributes.
-      if (action.payload.destinationIndex !== undefined) {
-        newState.destinationIndexNameExists = newState.indexNames.some(
-          name => newState.destinationIndex === name
-        );
-        newState.destinationIndexNameEmpty = newState.destinationIndex === '';
-        newState.destinationIndexNameValid = isValidIndexName(newState.destinationIndex);
-        newState.destinationIndexPatternTitleExists = newState.indexPatternTitles.some(
-          name => newState.destinationIndex === name
-        );
-      }
-
-      if (action.payload.indexNames !== undefined) {
-        newState.destinationIndexNameExists = newState.indexNames.some(
-          name => newState.destinationIndex === name
-        );
-        newState.sourceIndexNameExists = newState.indexNames.some(
-          name => newState.sourceIndex === name
-        );
-      }
-
-      if (action.payload.indexPatternTitles !== undefined) {
-        newState.destinationIndexPatternTitleExists = newState.indexPatternTitles.some(
-          name => newState.destinationIndex === name
-        );
-      }
-
-      if (action.payload.jobId !== undefined) {
-        newState.jobIdExists = newState.jobIds.some(id => newState.jobId === id);
-        newState.jobIdEmpty = newState.jobId === '';
-        newState.jobIdValid = isAnalyticsIdValid(newState.jobId);
-      }
-
-      if (action.payload.jobIds !== undefined) {
-        newState.jobIdExists = newState.jobIds.some(id => newState.jobId === id);
-      }
-
-      if (action.payload.sourceIndex !== undefined) {
-        newState.sourceIndexNameExists = newState.indexNames.some(
-          name => newState.sourceIndex === name
-        );
-        newState.sourceIndexNameEmpty = newState.sourceIndex === '';
-        newState.sourceIndexNameValid = isValidIndexName(newState.sourceIndex);
-      }
-
-      return validate(newState);
-  }
-
-  return state;
-}
-
 // List of system fields we want to ignore for the numeric field check.
 const OMIT_FIELDS: string[] = ['_source', '_type', '_index', '_id', '_version', '_score'];
 
-function getErrorMessage(error: any) {
+export function getErrorMessage(error: any) {
   if (typeof error === 'object' && typeof error.message === 'string') {
     return error.message;
   }
